requirements.txt
transformers==4.30.0
torch==2.0.1
streamlit==1.25.0
Pillow==10.0.0
sentence-transformers==2.2.2
python-dotenv==1.0.0